feat(key manage): display owner username or shared with usernames
